make uint64_t into size_t (same thing really) and remove a redundant cast

+++ src/Gpu/Device/Emulator/RuntimeManager/handleLaunchTask.cpp
@@ -15,7 +15,7 @@ namespace Gpu::Device {
 
         const auto& blob = kernels_.get(msg.kernelId);
         try {
-            threads_.insert(static_cast<uint32_t>(msg.kernelId),
+            threads_.insert(msg.kernelId,
                 std::jthread([blob, args] {
                     const ThreadRuntime rt(blob, args);
                     rt.run();
